requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
scipy>=1.10
opencv-python>=4.8
Pillow>=10.0
PyWavelets>=1.4
tqdm>=4.65
matplotlib>=3.7
